Move parser quickcheck "fuzz" to `nu-cmd-lang`
This should be faster to run with less command and to trigger failures
in the parser or basic const eval this should suffice. See the
`nu-parser` `parse_with_keywords` fuzz target for context.
