+++ crates/nu-cmd-lang/src/lib.rs
@@ -4,6 +4,8 @@ mod core_commands;
 mod default_context;
 pub mod example_support;
 mod example_test;
+#[cfg(test)]
+mod parse_const_test;
 
 pub use core_commands::*;
 pub use default_context::*;
